refactor: `InputDropDownViewModel` now directly inherits from `InputChoiceViewModel`
* Thanks to this, `InputSingleChoiceViewModel` could now be safely removed.
